move settingss to routes
