+++ app/routes/drawer/routes/settings/routes/generalAppInfo/index.js
